Revert "Move lazy_init under scipp folder again"
This reverts commit 7d4aede2d62a6bd38a6fcfa8147d59f5a06be577.
