+++ src/binding/python/openpmd_api/ScippLazyInit.py
@@ -1,7 +1,8 @@
 def series_to_scipp(series):
+
     import scipp
 
-    from . import DataLoader
+    from .scipp import DataLoader
 
     dl = DataLoader(series)
     return dl
